requirements.txt
playwright==1.49.0
pyyaml==6.0.2
python-dotenv==1.0.1
requests==2.32.3
